modified:   components/donate/DonateNav.jsx 	renamed:    components/donate/donorprofile/DonorProfile.jsx -> components/donorprofile/DonorProfile.jsx 	renamed:    components/donate/donorprofile/DonorProfileBody.jsx -> components/donorprofile/DonorProfileBody.jsx 	renamed:    components/donate/donorprofile/DonorProfileCard.jsx -> components/donorprofile/DonorProfileCard.jsx 	renamed:    pages/donate/profile/index.jsx -> pages/profile/index.jsx
