+++ components/donate/DonateNav.jsx
@@ -23,8 +23,8 @@ const DonateNav = () => {
       <div className="pb-[4px] hover:pb-0 hover:border-b-4 hover:border-black hover:text-black cursor-pointer transition p-2 rounded text-main font-satoshiMedium text-2xl">
         Earnings
       </div>
-      {pathname !== '/donate/profile' ? (
-        <Link href="/donate/profile">
+      {pathname !== '/profile' ? (
+        <Link href="/profile">
           <div className="pb-[4px] hover:pb-0 hover:border-b-4 hover:border-black hover:text-black cursor-pointer transition p-2 rounded text-main font-satoshiMedium text-2xl">
             Profile
           </div>
