+++ components/donorprofile/DonorProfileCard.jsx
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers'
 import Link from 'next/link'
-import { biobankNames } from '../../../constants/index'
+import { biobankNames } from '../../constants/index'
 
 const DonorProfileCard = ({ item }) => {
   const biobank = item.bioBank
@@ -9,7 +9,7 @@ const DonorProfileCard = ({ item }) => {
   const BioBankNames = biobankNames
 
   return (
-    <Link href={`/marketplace/nft?tokenId=${item.tokenId}`}>
+    <Link href={`/marketplace/nft?tokenId=${item.tokenId}&matchRating=0`}>
       <div className="w-fit min-w-[256px] max-w-[256px] bg-white flex flex-col px-5 pt-2 pb-4 mb-[2rem] mr-[2rem] drop-shadow-nftCard cursor-pointer rounded-2xl">
         <div className="py-3">
           <div className="flex flex-row justify-between mb-2">
